Improved binary buffer handling and fixed comm messages
